checkpoint refactoring
