+++ Cargo.toml
@@ -10,17 +10,17 @@ repository = "https://github.com/alexpasmantier/television"
 homepage = "https://github.com/alexpasmantier/television"
 keywords = ["search", "fuzzy", "preview", "tui", "terminal"]
 categories = [
-  "command-line-utilities",
-  "command-line-interface",
-  "concurrency",
-  "development-tools",
+    "command-line-utilities",
+    "command-line-interface",
+    "concurrency",
+    "development-tools",
 ]
 include = [
-  "LICENSE",
-  "README.md",
-  "crates/television/**/*.rs",
-  "build.rs",
-  ".config/config.toml",
+    "LICENSE",
+    "README.md",
+    "crates/television/**/*.rs",
+    "build.rs",
+    ".config/config.toml",
 ]
 rust-version = "1.80.0"
 
@@ -37,10 +37,10 @@ repository = "https://github.com/alexpasmantier/television"
 homepage = "https://github.com/alexpasmantier/television"
 keywords = ["search", "fuzzy", "preview", "tui", "terminal"]
 categories = [
-  "command-line-utilities",
-  "command-line-interface",
-  "concurrency",
-  "development-tools",
+    "command-line-utilities",
+    "command-line-interface",
+    "concurrency",
+    "development-tools",
 ]
 include = ["LICENSE", "README.md", "crates/television/**/*.rs", "build.rs"]
 rust-version = "1.80.0"
@@ -56,6 +56,7 @@ name = "tv"
 # workspace dependencies
 television-fuzzy = { path = "crates/television-fuzzy", version = "0.0.8" }
 television-derive = { path = "crates/television-derive", version = "0.0.8" }
+television-screen = { path = "crates/television-screen", version = "0.0.8" }
 television-channels = { path = "crates/television-channels", version = "0.0.8" }
 television-previewers = { path = "crates/television-previewers", version = "0.0.8" }
 television-utils = { path = "crates/television-utils", version = "0.0.8" }
