+++ crates/television-screen/Cargo.toml
@@ -1,0 +1,24 @@
+[package]
+name = "television-screen"
+version = "0.0.8"
+edition.workspace = true
+description.workspace = true
+license.workspace = true
+authors.workspace = true
+repository.workspace = true
+homepage.workspace = true
+keywords.workspace = true
+categories.workspace = true
+include.workspace = true
+rust-version.workspace = true
+readme.workspace = true
+
+[dependencies]
+ratatui = "0.29.0"
+serde = "1.0.215"
+television-utils = "0.0.8"
+television-channels = "0.0.8"
+television-previewers = "0.0.8"
+color-eyre = "0.6.3"
+ansi-to-tui = "7.0.0"
+syntect = "5.2.0"
