+++ crates/television-screen/src/mode.rs
@@ -1,5 +1,5 @@
-use crate::television::Mode;
 use ratatui::style::Color;
+use serde::{Deserialize, Serialize};
 
 const CHANNEL_COLOR: Color = Color::Indexed(222);
 const REMOTE_CONTROL_COLOR: Color = Color::Indexed(1);
@@ -11,4 +11,11 @@ pub fn mode_color(mode: Mode) -> Color {
         Mode::RemoteControl => REMOTE_CONTROL_COLOR,
         Mode::SendToChannel => SEND_TO_CHANNEL_COLOR,
     }
+}
+
+#[derive(PartialEq, Copy, Clone, Hash, Eq, Debug, Serialize, Deserialize)]
+pub enum Mode {
+    Channel,
+    RemoteControl,
+    SendToChannel,
 }
